WeightedLevenshtein ins/del weights.
Extend WeightedLevenshtein to have customizable insert / deletion weights.
Previously, insert / deletion weights were hardcoded at 1.0.  Customizing
them allows the caller to under-weight the insertion of a thin letter like
I or l to reflect the likelihood of OCR errors (for example).

This adds a new interface, CharacterInsDelInterface, which is an
adjunct to CharacterSubstitutionInterface.  The old behavior is preserved
if the caller does not provide a CharacterSubstitutionInterface subclass.

This also adds insert / deletion tests to the old
WeightedLevenshteinTest.testDistance, and adds a new
testDistanceCharacterInsDelInterface test.