Update Jaro-Winkler description in README
The previous readme made it same like Jaro-Winkler was the ideal typo detector,
when in actuality it is really only suited for typos caused by unsynchronized
high-speed typing between between both hands but does not account for actual
miskey errors such as hitting the wrong key altogether or advertently pressing
two keys instead of one. This is because Jaro-Winkler operates only on
transpositions and does not favorbly consider a string consisting strictly of
additions or permutitions with letters not already part of the word's alphabet
to be "similar" changes.